requirements.txt
asgiref==3.2.10
attrs==20.2.0
click==7.1.2
Django==3.1.2
django-extensions==3.0.9
factory-boy==3.1.0
Faker==4.14.0
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
packaging==20.4
Pillow==8.0.1
pluggy==0.13.1
psycopg2==2.8.6
py==1.9.0
pyparsing==2.4.7
pytest==6.1.2
pytest-django==4.1.0
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.1
Werkzeug==1.0.1
